Restructure styles within the app/frontend directory
Webpacker objects to multiple files with the same name in
app/frontend/packs, even if they have different file extensions.

The solution to this is to move application.scss up to the top-level
app/frontend directory and include it in the application.js pack.

This commit also includes movement of foundation stylesheets
and some colour palette changes, since this bugfix was
performed on a WIP branch.