Add inline_svg gem
This gem is used instead of just a helper method because the gem
handles webpacker path resolution, as well as IO or filepath
resolution, making it fit for a number of purposes.

The gem also handles accessibility, allowing the SVG to be titled,
and supports transforms, which we could use in the future instead of/
as well as CSS classes to rewrite SVG attributes.