Add system spec helper that will wait for a path to change
This helper is handy to use for complex AJAX interactions where
 simply waiting for a single request to complete may not be suffient.

A good example of this is ActiveSupport direct uploads, where there will be
an initial direct upload of the asset, followed by a XHR form
submission. Just waiting for AJAX to complete in this case is not sufficient
because the first request will complete and the test will advance without
waiting for the form submission request to fire.