Override DirectUploadsController from ActiveStorage to enforce attachment restrictions
Without this override, Rails will always upload the attachment, and will clean up after itself if the validations on a model run later on.

With this change, we check the direct upload params on their way in.
If the direct upload is being requested for a file that we know is going to be invalid, then we return an error immediately.