Modify Search to expect a single-value sort key instead of columns and direction
Separate sort keys and directions are suited to API designs, but not to
a traditional server-rendered application. The reason they are
unsuitable is that there's no native control that allows both a column
and direction to be submitted from a single input without javascript.

Using this pattern allows us to have a whitelist of permitted sorts,
while also allowing us to support unidirectional sort keys like "recent"
or "relevance".