Split sign folder button rendering logic out into it's own helper method
This helper now needs to consider states where the user is signed in or not, as
well as whether the sign has already been added to a folder or not.

Moving to a helper method allows it to be integration tested. I added it to
a helper module rather than the presenter as I felt it was a 'helper' to show
something _adjacent_ to a sign rather than a direct attribute on the sign.