Extend Administrate::Search to resolve some issues:
1. The query-table_name method assumes that the table name is the attribute pluralised.
This is almost immediately wrong for any association whose name is different from it's table
(e.g. Sign.contributor).

The solution is to look up the actual table name on the class_name being passed. If it is not present,
then we fall back to pluralising the attribute.

2. The tables_to_join method also assumes that the field in the dashboard matches the association.
To resolve this, a new option is looked at - `:source` (to match the ActiveRecord option of the same name.
This is used as the join condition if present, otherwise original behaviour is used.