Rules around whether the sign folders button can be rendered are more nuanced
We should not render anything if the sign is not in a state where it can be added to folders.

If it would normally be in a state to be added, but the user is not signed in, we should show a link to sign in.

If the user is signed in and it is in a state to be added, we can show the regular button.