Run jobs asynchrously during tests to avoid intermittent failures
When files are uploaded, they are processed. In particular, videos are
encoded. We need to disable this, as it is leading to completely random
test failures depending on whether the job finishes processing before
the test times out.

There may be some tests that require an inline adapter. In these
scenarios, we should just set the adapter for that session.