Fixing two bugs relating to video
Fixing two bugs, the first in javascript where jQuery was bieng called
from inline javascript and thats never going to work because inline
javascript CAN'T be defered till page load even if you put the `defer`
keyword in the declaration. Also changed it to not be a setinterval that
you need to cancel, but a setTimeout that you can ignore, since the
setTimeout was being called even if one was still in progress meanign
two calls happen at once.

The second is that the video encoder was encoding two copies of the same
video. Videos are reencoded to three variants upon upload, but the keys
used to identify the videos are generated from the video encoding
options. If those options are different, IE they are in a differnet
order, you end up with two identicle videos with difffernt keys.
Since the UI and the database thinks that the videos are already
encoded, it asks for the video, but the video backend uses the wrong
options to generate the key to look up the video and thinks it DOESN'T
exsit so it reencodes a new copy and meanwhile the UI gets back a string
saying "hold on just reencoding the video" when it expects a binary
video....