Removed that javascript document.onload code that changes the video (#353)
* Removed that javascript document.onload code that changes the video
elements after they are loaded and add the <i> overlay, prefering to add in the onclick and
onended events explicitly into the declaration of the video.
This means that any videos loaded after the page loads (IE When you hit
"Show More") will also get those onclick events etc

* removed the test for the video element having a controls attribute, as
in the old code the controls element was always removed eventually. We now default
to the controls not being present by default so this test makes no sense
anymore.

* Lets remove the test for controls being present on the video element,
since the controls used to be removed eventually. Nopw they default to
not being there in the first place.

* The signs id is unique, so lets use that instead of the sign.video.id
since freelexvideo signs don;t have a video method

* disable the specific eslint rule, not ALL rules, in video.js.
Remove a blank line in _video.html.erb

* commiting to ensure I don't loose code really. I have a more jqueryish
solution here, where the "onclick" event propagates up to document so the
event handler works but the "onended" event doesn't so that event
handler doesn't.

* remove some useless javascript from both video.js and the inline
javascript from the hero-unit

* removed the debugging, and simplified the function definitiona

* Update app/frontend/components/video.js

Co-authored-by: Gareth Jones <gareth.jones@ackama.com>

Co-authored-by: Jacob Briggs <jacob.briggs@ackama.com>
Co-authored-by: Gareth Jones <gareth.jones@ackama.com>