Update the dictionary if it is older than 1 month old when the application starts (#490)
* Update the dictionary if it is older than 1 month old when the application starts

This is more resilient than trying to intercept all process invocations. On heroku, the deploy failed due to assets:precompile being run for example.

Instead, we can download the database each time the app starts, to make sure we are using the most up-to-date version. Heroku will always update the most recent version, since the database is gitignored so not present in the slug.

* Update documentation to refer to the dictionary database being automatically downloaded