Sidekiq 7 updates version 2 (#593)
This PR re-adds the sidekiq upgrades with a couple of small changes that
should hopefully mean that thumbnail processing now works correctly. The
changes consist of a couple of lines in sidekiq batch. The issue was
that newer versions of redis are more strict with type checking, so we
can no longer send it boolean values (in the past it changed them to
strings behind the scenes). Because of this a couple of boolean trues
needed to be converted into strings. I have changed the smallest number
of lines that I needed to in order to get thumbnails generating as I am
not super familiar with the ins and outs of sidekiq batch. Because this
PR adds all of sidekiq batch back into the app so my changes would be
hard to find, I changed the following lines in
`vendor/gems/sidekiq-batch/lib/sidekiq/batch`:

- Line 251: pipeline.hset(batch_key, event_name, true) ->
pipeline.hset(batch_key, event_name, 'true')
- Line 295: cb_batch.callback_batch = true -> cb_batch.callback_batch =
'true'

Related: https://github.com/breamware/sidekiq-batch/issues/86

---------

Co-authored-by: Cameron Hopkinson <hopkincame@MSI>