NZSL-163: Display the dictionary database version in the application footer (#604)
This pull request adjusts the footer to display the user_version PRAGMA
value from the released dictionary database. This allows users to see
what version of the data is being used. This is especially handy to
check that the data has been updated, and to check what version of data
is used if an issue is noticed.


![image](https://github.com/ackama/nzsl-share/assets/292020/7dbce08a-da3e-4ff9-90d5-129c708f8cf9)