NZSL-160: Pull Signbank dictionary database from a public or private S3 location (#601)
This pull request modifies the boot process of NZSL Share to download a
file from a specified S3 location. This can be a public or private S3
object, and the URI is specified using an environment variable. We
attempt to download from the public location first, then attempt to
generate a presigned URL and download from that. If no keypair is
present and the public URL is not accessible (e.g. this is a prerelease
database), then the app continues to boot, but without a dictionary
database.

I'm expecting to provide a similar patch to
https://github.com/odnzsl/nzsl-online. We might consider extracting this
common code into a Rubygem at some point, since we fairly regularly end
up duplicating code.