Remove unused 'description' field from signs (#610)
This field was included in the signbank export, which is how we noticed
it was still there. It appears that it was
[added](https://github.com/ackama/nzsl-share/commit/87e78384725d024013acbec3e9a258fb7d0f1654)
to an early data modelling commit, and subsequently superseded by notes
and usage examples. It's not used in any view, dashboard, presenter or
anywhere else that I can see, so to avoid confusion, should be removed.