CC-2379: upgrade Rails to v7.2 (#731)
Update rails to 7.2

Update rspec-mocks to above 3.12.4 to solve " wrong number of arguments
(given 3, expected 0)" failures in specs

Fix all rubocop autocorrects

In `spec/presenters/sign_presenter_spec.rb` - Stubbed objects cannot
access the database, and activerecord needs a persisted object in order
to get a signed_id from it, so we need to change sign from a stubbed
object to a real one

Many changes to request specs to look at the http status returned,
rather than look for an exception

In `app/views/signs/edit/_attachment.html.erb` - Remove the `.sign` as
the `form.object` is already a `Sign`. Prior to the upgrade, it was a
`SignPresenter` and asking for the `.sign` would give you a `Sign`. Now,
for some reason, its not a `SignPresenter` anymore. I have validated
this change works manually, as well as via the tests.

After updating the rails defaults from 7.0 to 7.2, the various setup
tasks started failing (for example `bin/setup`) on
`require "active_storage/disk_controller"` in
`lib/active_storage_ext/app/controllers/disk_controller.rb`
I commented it out to see if it runs and it does and its because these
no longer can be manually required as they are autoloaded
https://guides.rubyonrails.org/upgrading_ruby_on_rails.html#autoloaded-paths-are-no-longer-in-$load-path