feat: update `@uppy` to v4 (#761)
The main change I had to do was with the `upload-start`(ed) event, which
now expects to be called with the full list of files and thus is not
appropriate to call within the uploading of each individual file.

While the end fix seems to be straightforward, it was made a lot harder
to track down by the fact we were settling the promises meaning the
error that got thrown after just blindly switching to `upload-start`
without moving the calls was silently eaten, and then the postprocesser
would error instead since it assumes the upload(s) were successful.

To make it easier to catch, I've switched us to using `Promise.all`
which will reject if any of the promises reject, and that seems to be
handled properly by Uppy (i.e. it does not call the postprocessor).

I have also updated the import of `EventTracker` to be `EventManager`
because it will be moved to `@uppy/core` in v5 with the new name only.

Addresses GHSA-mwcw-c2x4-8c55